+++ R/Validator.R
@@ -136,6 +136,7 @@ Validator <- R6::R6Class(
         check_affiliation = check_affiliation,
         check_credit = check_credit,
         check_coi = check_coi,
+        check_coi_column_rename = check_coi_column_rename,
         check_author_acknowledgee_values = check_author_acknowledgee_values,
         check_corresponding_non_author = check_corresponding_non_author,
         check_missing_author_acknowledgee = check_missing_author_acknowledgee,