+++ NEWS.md
@@ -1,3 +1,9 @@
+# tenzing 0.4.1
+
+## Bug fixes
+- Fixed a rendering bug with the spreadsheet view module
+- Meaningful error message regarding "Conflict of interest" - "Declares" column renaming
+
 # tenzing 0.4.0
 
 ## New features